requirements/auth_requirements.txt
asgiref==3.3.1
autopep8==1.6.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
Django==3.1.7
django-cors-headers==3.7.0
django-redis==4.12.1
djangorestframework==3.13.1
gunicorn==20.0.4
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
oauthlib==3.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.17
protobuf==3.19.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.8.1
python-dotenv==0.19.2
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==4.3.3
urllib3==1.26.4
wcwidth==0.2.5
